requirements.txt
opencv-python>=4.5.5
numpy>=1.21.0
ultralytics>=8.0.0
easyocr>=1.6.0
pandas>=1.3.0
scikit-learn>=1.0.0
requests>=2.28.0
Flask>=2.0.0
python-dotenv>=0.19.0 